Extract the API store into a separate package: ecamp/hal-json-vuex
