+++ frontend/src/main.js
@@ -1,10 +1,9 @@
 import Vue from 'vue'
 import App from './App.vue'
 import router from '@/router'
-import { vuetify, auth, apiStore, store, filterLoading, formBaseComponents, ignoreNativeBindingWarnMessages, i18n, veeValidate, vueMoment, moment } from '@/plugins'
+import { vuetify, auth, store, filterLoading, formBaseComponents, ignoreNativeBindingWarnMessages, i18n, veeValidate, vueMoment, moment } from './plugins'
 
 Vue.use(auth)
-Vue.use(apiStore)
 Vue.use(filterLoading)
 Vue.use(formBaseComponents)
 Vue.use(ignoreNativeBindingWarnMessages)
