--- frontend/src/plugins/store/__tests__/.eslintrc.js
@@ -1,5 +1,0 @@
-module.exports = {
-  env: {
-    jest: true
-  }
-}
