--- frontend/src/plugins/store/__tests__/resources/collection-complete.json
@@ -1,158 +1,0 @@
-{
-  "serverResponse": {
-    "_page": 0,
-    "_per_page": -1,
-    "_total": 3,
-    "page_count": 1,
-    "_links": {
-      "self": {
-        "href": "/camps/1/activities"
-      },
-      "first": {
-        "href": "/camps/1/activities?page=0"
-      },
-      "next": {
-        "href": "/camps/1/activities?page=1"
-      }
-    },
-    "_embedded": {
-      "items": [
-        {
-          "id": 2394,
-          "title": "Volleyball",
-          "_embedded": {
-            "type": {
-              "id": 8877,
-              "name": "LS",
-              "color": "#00aa00",
-              "_links": {
-                "self": {
-                  "href": "/activityTypes/8877"
-                }
-              }
-            }
-          },
-          "_links": {
-            "self": {
-              "href": "/activities/2394"
-            }
-          }
-        },
-        {
-          "id": 2362,
-          "title": "Geländespiel",
-          "_embedded": {
-            "type": {
-              "id": 8879,
-              "name": "LP",
-              "color": "#aaaa00",
-              "_links": {
-                "self": {
-                  "href": "/activityTypes/8879"
-                }
-              }
-            }
-          },
-          "_links": {
-            "self": {
-              "href": "/activities/2362"
-            }
-          }
-        },
-        {
-          "id": 2402,
-          "title": "TABS",
-          "type": {
-            "id": 8879,
-            "name": "LP",
-            "color": "#aaaa00",
-            "_links": {
-              "self": {
-                "href": "/activityTypes/8879"
-              }
-            }
-          },
-          "_links": {
-            "self": {
-              "href": "/activities/2402"
-            }
-          }
-        }
-      ]
-    }
-  },
-  "storeState": {
-    "/camps/1/activities": {
-      "_meta": {
-        "self": "/camps/1/activities"
-      },
-      "_page": 0,
-      "_per_page": -1,
-      "_total": 3,
-      "page_count": 2,
-      "first": {
-        "href": "/camps/1/activities?page=0"
-      },
-      "next": {
-        "href": "/camps/1/activities?page=1"
-      },
-      "items": [
-        {
-          "href": "/activities/2394"
-        },
-        {
-          "href": "/activities/2362"
-        },
-        {
-          "href": "/activities/2402"
-        }
-      ]
-    },
-    "/activityTypes/8877": {
-      "id": 8877,
-      "color": "#00aa00",
-      "name": "LS",
-      "_meta": {
-        "self": "/activityTypes/8877"
-      }
-    },
-    "/activityTypes/8879": {
-      "id": 8879,
-      "color": "#aaaa00",
-      "name": "LP",
-      "_meta": {
-        "self": "/activityTypes/8879"
-      }
-    },
-    "/activities/2362": {
-      "id": 2362,
-      "type": {
-        "href": "/activityTypes/8879"
-      },
-      "_meta": {
-        "self": "/activities/2362"
-      },
-      "title": "Geländespiel"
-    },
-    "/activities/2394": {
-      "id": 2394,
-      "type": {
-        "href": "/activityTypes/8877"
-      },
-      "_meta": {
-        "self": "/activities/2394"
-      },
-      "title": "Volleyball"
-    },
-    "/activities/2402": {
-      "id": 2402,
-      "type": {
-        "href": "/activityTypes/8879"
-      },
-      "_meta": {
-        "self": "/activities/2402"
-      },
-      "title": "TABS"
-    }
-  }
-}
