--- frontend/src/plugins/store/__tests__/resources/collection-page1.json
@@ -1,84 +1,0 @@
-{
-  "serverResponse": {
-    "_page": 1,
-    "_per_page": 2,
-    "_total": 3,
-    "page_count": 2,
-    "_links": {
-      "self": {
-        "href": "/camps/1/activities?page=1&page_size=2"
-      },
-      "first": {
-        "href": "/camps/1/activities?page=0&page_size=2"
-      },
-      "prev": {
-        "href": "/camps/1/activities?page=0&page_size=2"
-      }
-    },
-    "_embedded": {
-      "items": [
-        {
-          "id": 2402,
-          "title": "TABS",
-          "_embedded": {
-            "type": {
-              "id": 8879,
-              "name": "LP",
-              "color": "#aaaa00",
-              "_links": {
-                "self": {
-                  "href": "/activityTypes/8879"
-                }
-              }
-            }
-          },
-          "_links": {
-            "self": {
-              "href": "/activities/2402"
-            }
-          }
-        }
-      ]
-    }
-  },
-  "storeState": {
-    "/camps/1/activities?page=1&page_size=2": {
-      "_meta": {
-        "self": "/camps/1/activities?page=1&page_size=2"
-      },
-      "first": {
-        "href": "/camps/1/activities?page=0&page_size=2"
-      },
-      "prev": {
-        "href": "/camps/1/activities?page=0&page_size=2"
-      },
-      "items": [
-        {
-          "href": "/activities/2402"
-        }
-      ],
-      "_page": 1,
-      "_per_page": 2,
-      "_total": 3,
-      "page_count": 2
-    },
-    "/activityTypes/8879": {
-      "id": 8879,
-      "color": "#aaaa00",
-      "name": "LP",
-      "_meta": {
-        "self": "/activityTypes/8879"
-      }
-    },
-    "/activities/2402": {
-      "id": 2402,
-      "_meta": {
-        "self": "/activities/2402"
-      },
-      "title": "TABS",
-      "type": {
-        "href": "/activityTypes/8879"
-      }
-    }
-  }
-}
