--- frontend/src/plugins/store/__tests__/resources/embedded-collection.json
@@ -1,90 +1,0 @@
-{
-  "serverResponse": {
-    "id": 1,
-    "_embedded": {
-      "periods": [
-        {
-          "id": 104,
-          "start": "01-01-2019",
-          "end": "03-01-2019",
-          "_links": {
-            "self": {
-              "href": "/periods/104"
-            },
-            "camp": {
-              "href": "/camps/1"
-            },
-            "activities": {
-              "href": "/periods/104/activities"
-            }
-          }
-        },
-        {
-          "id": 128,
-          "start": "12-04-2019",
-          "end": "19-04-2019",
-          "_links": {
-            "self": {
-              "href": "/periods/128"
-            },
-            "camp": {
-              "href": "/camps/1"
-            },
-            "activities": {
-              "href": "/periods/128/activities"
-            }
-          }
-        }
-      ]
-    },
-    "_links": {
-      "self": {
-        "href": "/camps/1"
-      }
-    }
-  },
-  "storeState": {
-    "/periods/104": {
-      "id": 104,
-      "start": "01-01-2019",
-      "end": "03-01-2019",
-      "camp": {
-        "href": "/camps/1"
-      },
-      "activities": {
-        "href": "/periods/104/activities"
-      },
-      "_meta": {
-        "self": "/periods/104"
-      }
-    },
-    "/periods/128": {
-      "id": 128,
-      "start": "12-04-2019",
-      "end": "19-04-2019",
-      "camp": {
-        "href": "/camps/1"
-      },
-      "activities": {
-        "href": "/periods/128/activities"
-      },
-      "_meta": {
-        "self": "/periods/128"
-      }
-    },
-    "/camps/1": {
-      "id": 1,
-      "periods": [
-        {
-          "href": "/periods/104"
-        },
-        {
-          "href": "/periods/128"
-        }
-      ],
-      "_meta": {
-        "self": "/camps/1"
-      }
-    }
-  }
-}
