--- frontend/src/plugins/store/__tests__/resources/embedded-single-entity.json
@@ -1,43 +1,0 @@
-{
-  "serverResponse": {
-    "_embedded": {
-      "campType": {
-        "id": 20,
-        "name": "camp",
-        "js": true,
-        "targetGroup": "Kids",
-        "_links": {
-          "self": {
-            "href": "/campTypes/20"
-          }
-        }
-      }
-    },
-    "id": 1,
-    "_links": {
-      "self": {
-        "href": "/camps/1"
-      }
-    }
-  },
-  "storeState": {
-    "/campTypes/20": {
-      "id": 20,
-      "name": "camp",
-      "js": true,
-      "targetGroup": "Kids",
-      "_meta": {
-        "self": "/campTypes/20"
-      }
-    },
-    "/camps/1": {
-      "campType": {
-        "href": "/campTypes/20"
-      },
-      "id": 1,
-      "_meta": {
-        "self": "/camps/1"
-      }
-    }
-  }
-}
