--- frontend/src/plugins/store/__tests__/resources/linked-collection.json
@@ -1,24 +1,0 @@
-{
-  "serverResponse": {
-    "id": 1,
-    "_links": {
-      "self": {
-        "href": "/camps/1"
-      },
-      "activities": {
-        "href": "/camps/1/activities"
-      }
-    }
-  },
-  "storeState": {
-    "/camps/1": {
-      "id": 1,
-      "activities": {
-        "href": "/camps/1/activities"
-      },
-      "_meta": {
-        "self": "/camps/1"
-      }
-    }
-  }
-}
