--- frontend/src/plugins/store/__tests__/resources/multiple-references-to-user.json
@@ -1,35 +1,0 @@
-{
-  "id": 99,
-  "name": "Pfadi Züri",
-  "_embedded": {
-    "users": [
-      {
-        "id": 1,
-        "name": "Linux",
-        "_embedded": {
-          "lastViewedCamp": {
-            "id": 123,
-            "_links": {
-              "owner": {
-                "href": "/users/1"
-              },
-              "self": {
-                "href": "/camps/123"
-              }
-            }
-          }
-        },
-        "_links": {
-          "self": {
-            "href": "/users/1"
-          }
-        }
-      }
-    ]
-  },
-  "_links": {
-    "self": {
-      "href": "/groups/99"
-    }
-  }
-}
