--- frontend/src/plugins/store/__tests__/resources/reference-to-single-entity.json
@@ -1,30 +1,0 @@
-{
-  "serverResponse": {
-    "_embedded": {
-      "campType": {
-        "_links": {
-          "self": {
-            "href": "/campTypes/20"
-          }
-        }
-      }
-    },
-    "id": 1,
-    "_links": {
-      "self": {
-        "href": "/camps/1"
-      }
-    }
-  },
-  "storeState": {
-    "/camps/1": {
-      "campType": {
-        "href": "/campTypes/20"
-      },
-      "id": 1,
-      "_meta": {
-        "self": "/camps/1"
-      }
-    }
-  }
-}
