--- frontend/src/plugins/store/__tests__/resources/templated-link.json
@@ -1,54 +1,0 @@
-{
-  "linkingServerResponse": {
-    "id": 1,
-    "_links": {
-      "self": {
-        "href": "/camps/1"
-      },
-      "users": {
-        "href": "http://localhost/camps/1/users{/id}",
-        "templated": true
-      }
-    }
-  },
-  "linkedServerResponse": {
-    "id": 83,
-    "name": "Pflock",
-    "_links": {
-      "self": {
-        "href": "/camps/1/users/83"
-      }
-    }
-  },
-  "storeStateBeforeLinkedLoaded": {
-    "/camps/1": {
-      "id": 1,
-      "users": {
-        "href": "http://localhost/camps/1/users{/id}",
-        "templated": true
-      },
-      "_meta": {
-        "self": "/camps/1"
-      }
-    }
-  },
-  "storeStateAfterLinkedLoaded": {
-    "/camps/1": {
-      "id": 1,
-      "users": {
-        "href": "http://localhost/camps/1/users{/id}",
-        "templated": true
-      },
-      "_meta": {
-        "self": "/camps/1"
-      }
-    },
-    "/camps/1/users/83": {
-      "id": 83,
-      "name": "Pflock",
-      "_meta": {
-        "self": "/camps/1/users/83"
-      }
-    }
-  }
-}
